backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
groq>=0.9.0
langgraph>=1.1.3
langchain-core>=1.2.23
langchain>=1.2.13
sqlalchemy>=2.0.31
psycopg2-binary>=2.9.9
pymysql>=1.1.1
duckdb>=0.10.3
pandas>=2.2.2
openpyxl>=3.1.5
python-multipart>=0.0.9
pydantic>=2.11.0
pytest>=8.2.2
pytest-asyncio>=0.23.7
httpx>=0.27.0
sqlalchemy-libsql>=0.1.0
python-dotenv>=1.0.1
